+++ changelog.md
@@ -6,6 +6,7 @@
 * Add configurable poll timeout
 * Add special error code for message not received within poll time
 * Add topic field for messages
+* Add commit register for periodic commit
 
 <a name="0.9.1"></a>
 ## 0.9.1 (2017-04-19)